Refactoring done
No import loop.
To be tested
