+++ Classes/NetworkEnergy.py
@@ -26,7 +26,7 @@ import os.path
 import json
 
 import Domoticz
-from Modules.output import sendZigateCmd, maskChannel
+from Modules.basicOutputs import sendZigateCmd, maskChannel
 from Classes.AdminWidgets import AdminWidgets
 
 CHANNELS = [ '11', '12', '13','14','15','16','17','18','19','20','21','22','23','24','25','26']
