+++ Modules/bindings.py
@@ -5,8 +5,9 @@
 #
 
 import Domoticz
-import Modules.output
 from time import time
+
+from Modules.basicOutputs import sendZigateCmd
 from Modules.logging import loggingOutput
 from Modules.zigateConsts import CLUSTERS_LIST
 
@@ -287,32 +288,26 @@ def iterable(obj):
     else:
         return True
 
-def callBackForWebBindIfNeeded( self , srcNWKID ):
 
-    """
-    Check that WebBind are well set
-    """
 
-    if srcNWKID not in self.ListOfDevices:
-        return
-    if 'WebBind' not in self.ListOfDevices[srcNWKID]:
-        return
 
-    for Ep in list(self.ListOfDevices[srcNWKID]['WebBind']):
-        for ClusterId in list(self.ListOfDevices[srcNWKID]['WebBind'][ Ep ]):
-            for destNwkid in list(self.ListOfDevices[srcNWKID]['WebBind'][ Ep ][ClusterId]):
-                if destNwkid in ('Stamp','Target','TargetIEEE','SourceIEEE','TargetEp','Phase','Status'):
-                    Domoticz.Error("---> delete  destNwkid: %s" %( destNwkid))
-                    del self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]
-                elif ('Phase' in self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid] and \
-                                 self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['Phase'] == 'requested'):
-                    if ('Stamp' in self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid] and time() < self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['Stamp']+ 5):    # Let's wait 5s before trying again
-                        continue
-                    loggingOutput( self, 'Log', "Redo a WebBind for device %s" %(srcNWKID))
-                    sourceIeee = self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['SourceIEEE']
-                    destIeee = self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['TargetIEEE']
-                    destEp = self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['TargetEp']
-                    # Perforning the bind
-                    webBind(self, sourceIeee, Ep, destIeee, destEp, ClusterId)
 
-                    self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['Stamp'] = int(time())
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
