+++ Modules/checkingUpdate.py
@@ -64,7 +64,7 @@ def checkPluginUpdate( currentVersion, availVersion):
 
 def checkFirmwareUpdate( currentMajorVersion, currentFirmwareVersion, availfirmMajor, availfirmMinor):
 
-    if not availfirmMinor or not currentFirmwareVersion:
+    if not (availfirmMinor and currentFirmwareVersion):
         return False
     if int(availfirmMinor,16) > int(currentFirmwareVersion,16):
         Domoticz.Debug("checkFirmwareUpdate - Firmware update available")
