+++ Modules/pairingProcess.py
@@ -21,19 +21,18 @@ from Modules.actuators import actuators
 from Modules.schneider_wiser import schneider_wiser_registration
 from Modules.legrand_netatmo import legrand_fc01
 from Modules.bindings import unbindDevice, bindDevice, rebind_Clusters
-from Modules.output import  sendZigateCmd,  \
-        identifyEffect, setXiaomiVibrationSensitivity, \
-        getListofAttribute, \
-        setPowerOn_OnOff, \
-        scene_membership_request, \
-        ReadAttributeRequest_0000_basic, \
+from Modules.basicOutputs import  sendZigateCmd, identifyEffect, getListofAttribute
+
+        
+from Modules.readAttributes import ReadAttributeRequest_0000_basic, \
         ReadAttributeRequest_0000, ReadAttributeRequest_0001, ReadAttributeRequest_0006, ReadAttributeRequest_0008, \
         ReadAttributeRequest_0100, \
         ReadAttributeRequest_000C, ReadAttributeRequest_0102, ReadAttributeRequest_0201, ReadAttributeRequest_0204, ReadAttributeRequest_0300,  \
         ReadAttributeRequest_0400, ReadAttributeRequest_0402, ReadAttributeRequest_0403, ReadAttributeRequest_0405, \
-        ReadAttributeRequest_0406, ReadAttributeRequest_0500, ReadAttributeRequest_0502, ReadAttributeRequest_0702, ReadAttributeRequest_000f, ReadAttributeRequest_fc01, ReadAttributeRequest_fc21
+        ReadAttributeRequest_0406, ReadAttributeRequest_0500, ReadAttributeRequest_0502, ReadAttributeRequest_0702, ReadAttributeRequest_000f, \
+        ReadAttributeRequest_fc01, ReadAttributeRequest_fc21
 
-from Modules.lumi import enableOppleSwitch
+from Modules.lumi import enableOppleSwitch, setXiaomiVibrationSensitivity
 from Modules.livolo import livolo_bind
 from Modules.orvibo import OrviboRegistration
 from Modules.configureReporting import processConfigureReporting
